cập nhật phần phân quyền truy cập trang

+++ server/modules/super-admin/component/component.service.js
@@ -5,10 +5,10 @@ const { Privilege, Role, Link, Component } = require('../../../models').schema;
  * @id id của công ty
  */
 exports.getComponents = async (company, query) => {
-    var page = query.page;
-    var limit = query.limit;
-    var currentRole = query.currentRole;
-    var linkId = query.linkId;
+    let page = query.page;
+    let limit = query.limit;
+    let currentRole = query.currentRole;
+    let linkId = query.linkId;
     
     if (!page && !limit && !currentRole && !linkId){
         return await Component
@@ -18,10 +18,10 @@ exports.getComponents = async (company, query) => {
                 { path: 'link', model: Link },
             ]);
     } else if (page && limit && !currentRole && !linkId) {
-        const option = (query.key && query.value)
+        let option = (query.key && query.value)
             ? Object.assign({company}, {[`${query.key}`]: new RegExp(query.value, "i")})
             : {company};
-        console.log("option: ", option);
+        
         return await Component
             .paginate( option , { 
                 page, 
@@ -32,22 +32,22 @@ exports.getComponents = async (company, query) => {
                 ]
             });
     } else if (!page && !limit && currentRole && linkId) {
-        const role = await Role.findById(currentRole);
+        let role = await Role.findById(currentRole);
         let roleArr = [role._id];
         roleArr = roleArr.concat(role.parents);
         
-        const link = await Link.findById(linkId)
+        let link = await Link.findById(linkId)
             .populate([
                 { path: 'components', model: Component }
             ]);
             
-        const data = await Privilege.find({
+        let data = await Privilege.find({
             roleId: { $in: roleArr },
             resourceType: 'Component',
             resourceId: { $in: link.components }
         }).distinct('resourceId');
 
-        const components = await Component.find({ _id: { $in: data } });
+        let components = await Component.find({ _id: { $in: data } });
 
         return components;
     }
@@ -71,7 +71,7 @@ exports.getComponent = async (id) => {
  * @data dữ liệu component
  */
 exports.createComponent = async(data) => {
-    const check = await Component.findOne({name: data.name});
+    let check = await Component.findOne({name: data.name});
 
     if(check) {
         throw ['component_name_exist'];
@@ -80,7 +80,8 @@ exports.createComponent = async(data) => {
     return await Component.create({
         name: data.name,
         description: data.description,
-        company: data.company
+        company: data.company,
+        deleteSoft: false
     });
 }
 
@@ -90,8 +91,7 @@ exports.createComponent = async(data) => {
  * @data dữ liệu
  */
 exports.editComponent = async(id, data) => {
-    console.log("data component: ", data)
-    const component = await Component
+    let component = await Component
         .findById(id)
         .populate({ path: 'roles', model: Privilege, populate: {path: 'roleId', model: Role } });
 
@@ -108,14 +108,20 @@ exports.editComponent = async(id, data) => {
  * Xóa component
  * @id id component
  */
-exports.deleteComponent = async(id) => {
-    const relationshiopDelete = await Privilege.deleteMany({
-        resourceId: id,
-        resourceType: 'Component'
-    });
-    const deleteComponent = await Component.deleteOne({ _id: id });
+exports.deleteComponent = async(id, type) => {
+    if(type === 'soft'){
+        let component = await Component.findById(id);
+        component.deleteSoft = true;
+        await component.save();
+    } else {
+        await Privilege.deleteMany({
+            resourceId: id,
+            resourceType: 'Component'
+        });
+        await Component.deleteOne({ _id: id });
+    }
 
-    return {relationshiopDelete, deleteComponent};
+    return id;
 }
 
 /**
@@ -129,7 +135,7 @@ exports.relationshipComponentRole = async(componentId, roleArr) => {
         resourceType: 'Component'
     });
 
-    const data = roleArr.map( role => {
+    let data = roleArr.map( role => {
         return {
             resourceId: componentId,
             resourceType: 'Component',
@@ -137,7 +143,7 @@ exports.relationshipComponentRole = async(componentId, roleArr) => {
         };
     });
 
-    const privilege = await Privilege.insertMany(data);
+    let privilege = await Privilege.insertMany(data);
 
     return privilege;
 }
